requirements.txt
pyTelegramBotAPIpyTelegramBotAPI
python-dotenv
